pyproject.toml
[project]
name = "assistant"
version = "0.1.0"
description = "An AI agent that organizes your knowledge from multiple sources and summarizes topics"
readme = "README.md"
requires-python = ">=3.13"
license = { file = "LICENSE" }
authors = [
    { name = "Filippo Pacifici" }
]
dependencies = [
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pyyaml>=6.0",
    "evernote-backup>=1.13.1",
    "keyring==25.7.0",
    "langchain-community>=0.4.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langchain_core>=1.2.16",
    "langchain-postgres>=0.0.17",
    "python-dotenv>=1.2.1",
    "langgraph-checkpoint-postgres>=3.0.4",
    "textual>=8.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "ruff>=0.1.6",
    "pre-commit>=3.5.0",
    "bandit[toml]>=1.7.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/assistant"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=assistant --cov-report=term-missing --cov-report=html"

[tool.mypy]
python_version = "3.13"
mypy_path = "src"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_concatenate = true

[[tool.mypy.overrides]]
module = ["yaml", "yaml.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py313"
line-length = 90

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
    "B",   # flake8-bugbear
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "ICN", # flake8-import-conventions
    "G",   # flake8-logging-format
    "PIE", # flake8-pie
    "T20", # flake8-print
    "PT",  # flake8-pytest-style
    "Q",   # flake8-quotes
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate
    "PD",  # pandas-vet
    "PL",  # pylint
    "TRY", # tryceratops
    "NPY", # numpy-specific rules
    "RUF", # ruff-specific rules
]
ignore = [
    "COM812", # Trailing comma rule conflicts with formatter
    "ISC001", # Implicit string concatenation conflicts with formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["PLR2004", "S101", "TC001", "TC002", "TC003"]
"src/assistant/export.py" = ["E402", "TC001", "TC002", "TC003"]
"src/assistant/restore.py" = ["E402", "TC001", "TC002", "TC003"]
"src/assistant/models/content.py" = ["E402", "TC001", "TC002", "TC003"]
"src/assistant/models/database.py" = ["PLC0415"]

[tool.pyright]
extraPaths = ["src"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", ".venv", "build", "dist"]
skips = ["B101"]  # Skip assert_used in test files
